Move plotting tests into own file
